pyproject.toml
[tool.poetry]
name = "soc-network"
version = "0.1.0"
description = ""
authors = ["rudich <rudi4.clarinet@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
python-dotenv = "^1.0.1"
dj-database-url = "^2.3.0"
markdown = "^3.7"
django-filter = "^24.3"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
djangorestframework-simplejwt = "^5.3.1"
django-extensions = "^3.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
